Adrian/network component

+++ lib/flowEngine/src/Renderer.cpp
@@ -139,7 +139,7 @@ namespace flow
 
         // flip the output fiew for the shader pass
         sf::View view = mWindowRef->getView();
-        view.setSize({ view.getSize().x, -fabs(view.getSize().y) });
+        view.setSize({ view.getSize().x, static_cast<float>(-fabs(view.getSize().y)) });
         mWindowRef->setView(view);
 
         mWindowRef->draw(sf::Sprite(output->getTexture())); // Output the final composite
